trunc and padding functions
